build(Cargo.toml): 2021 edition, dependency bumps
Thanks for this handy crate! I used it for a lil' script to further
bloat my `$PS1` with info about my Starlink connection because I love
cramming it full of questionably useful information like my indoor air
quality/heat pump state/satellite doohickeys.

In return, I give you this changeset of somewhat dubious utility, seeing
as your crate is functionally identical without it. Take it or leave it,
I guess, 'cuz it's free either way.

- Major version bump to 1.0.0 since this could theoretically break
someone's CI if they are using a museumworthy toolchain (idk if this is
idiomatic for edition bumps, I'm shooting from the hip here)

- 2018 -> 2021 because the dep updates appeared to generate code relying
on `TryInto`'s prelude presence; it may be possible to fix this without
an edition upgrade but I didn't investigate this on account of I need to
get back to solving heinous JVM classpath issues at the job that pays
for my existence

- bumped tonic & tonic-build from 0.8 -> 0.10.2

- bumped prost from 0.11 -> 0.12.3

- bumped prost-types from 0.11 -> 0.12.3

- bumped async-stream from 0.3 to 0.3.5

- bumped tokio from 1.5 to 1.34.0, added `macros` feature to fix
resulting compilation errors

I tested this locally with the aforementioned script, but as that's just
a `GetStatusRequest` the prudent thing to do would be to also try out
your examples... but I'm lazy, so, yeah. You are obviously welcome to do
anything or nothing with this PR as you please.